+++ app/assets/javascripts/folio/console/base.js
@@ -13,6 +13,7 @@
 //= require js-cookie/src/js.cookie
 //= require cleave.js/dist/cleave
 //= require jquery.kinetic/index
+//= require dropzone/dist/dropzone
 
 //= require jquery-ui/jquery-ui
 //= require selectize/dist/js/standalone/selectize
@@ -66,6 +67,7 @@
 //= require folio/console/modules/simple-form-with-atoms
 //= require folio/console/nested_model_controls/nested_model_controls
 //= require folio/console/new_record_modal/new_record_modal
+//= require folio/console/private_attachments/single_dropzone/single_dropzone
 //= require folio/console/publishable_inputs/publishable_inputs
 //= require folio/console/react_picker/react_picker
 //= require folio/console/single_file_select/single_file_select